[async] General cleanup
